od: refactor: split into modules
